clean up + fixture, FakeApi and onlyWithSandboxAp
