+++ tests/Pest.php
@@ -1,39 +1,19 @@
 <?php
 
-use GuzzleHttp\Promise\PromiseInterface;
 use TioJobs\AsaasPhp\Core\Asaas;
+use TioJobs\AsaasPhp\Tests\TestCase;
+use GuzzleHttp\Promise\PromiseInterface;
 
-uses(\TioJobs\AsaasPhp\Tests\TestCase::class)->in(__DIR__);
+uses(
+    TestCase::class,
+    Illuminate\Foundation\Testing\RefreshDatabase::class,
+)->in('Feature');
 
-function generateCustomer(): array
-{
-    $faker = \Faker\Factory::create('pt_BR');
-
-    return [
-        'name' => $faker->name,
-        'cpfCnpj' => $faker->cpf,
-        'email' => $faker->safeEmail,
-        'phone' => '16'.$faker->landline,
-        'mobilePhone' => '16'.$faker->cellphone,
-        'address' => explode(',', $faker->address)[1],
-        'addressNumber' => $faker->randomNumber(4),
-        'complement' => '',
-        'province' => $faker->city,
-        'postalCode' => $faker->postcode,
-        'externalReference' => '',
-        'notificationDisabled' => false,
-        'additionalEmails' => '',
-        'municipalInscription' => '',
-        'stateInscription' => '',
-        'observations' => 'Generated by faker test',
-        'groupName' => 'Tests',
-        'company' => $faker->company,
-    ];
-}
+require_once('Concerns/Factories.php');
 
 function asaasPhp(?string $apiKey = null, string $mode = 'sandbox'): Asaas
 {
-    if(config('asaas-php.allow_sub_accounts')){
+    if(config('asaas-php.allow_sub_accounts')) {
         return new Asaas((string) $apiKey, $mode);
     }
 
@@ -45,28 +25,21 @@ function fakeResponse(array|null|string $body = null, int $status = 200, array $
     return asaasPhp()->response($body, $status, $headers);
 }
 
-function generateCompany(): array
+function fixture(string $path): array
 {
-    $faker = \Faker\Factory::create('pt_BR');
+    $path = str_replace(['/', '.php'], [DIRECTORY_SEPARATOR, ''], __DIR__ . '/Fixtures/' . $path) . '.php';
 
-    return [
-        'name' => $faker->name,
-        'cpfCnpj' => $faker->cnpj,
-        'email' => $faker->safeEmail,
-        'phone' => '16'.$faker->landline,
-        'mobilePhone' => '16'.$faker->cellphone,
-        'address' => 'Rua Visconde do Rio Branco',
-        'addressNumber' => $faker->randomNumber(4),
-        'complement' => '',
-        'province' => 'Ribeirão Preto',
-        'postalCode' => '14015-000',
-        'externalReference' => '',
-        'notificationDisabled' => false,
-        'additionalEmails' => '',
-        'municipalInscription' => '',
-        'stateInscription' => '',
-        'observations' => 'Generated by faker test',
-        'groupName' => 'Tests',
-        'company' => $faker->company,
-    ];
+    return include $path;
+}
+
+function onlyWithSandboxApi()
+{
+    if (FakeApi()) {
+        test()->markTestSkipped('Only runs with Asaas API [ACCESS_ASAAS_API=true]');
+    }
+}
+
+function FakeApi(): bool
+{
+    return (bool) env('ACCESS_ASAAS_API') === false;
 }
